Add card show page
Provides detailed information about a specific card, including the
description which is omitted from the table view. Adds links to these
new pages to the card names. Adds link to NPC page from NPC sources.
Card sources are now returned as an array which can be more properly
displayed in a list.