Add card selection
Adds a new page where the user can set up or update their deck, using an
interface similar to the in-game list. Users can compare the two side by
side to see which cards they are missing between the app and the game.
Also adds the option to hide owned cards on the My Cards page.