Add proper card sources
Non-NPC/pack card sources are now stored with a proper model instead of
a comma-separated string. This allows the sources to be properly searched
in the cards search form. Some special handling is done for NPC/pack
sources to allow them to be part of the form. This update requires the
sources task to be run again.