Fix card select navigation warning
The listener continues to stick around between page loads because of the
nature of Turbolinks, so we check that we're still on the right page
before showing the alert. We make sure that any old listeners are
unregistered when navigating to the page again in order to avoid
Dixie Kong's Double Trouble.