Update NPC/Card tasks
Updates the NPC task to pull the Level data from the API instead of
relying on the CSV which can become out of date. Unfortunately this
requires scraping the entire endpoint, so this will take a little
longer. Fixes an issue with both tasks where an exception would be
thrown when trying to look up a missing card/NPC.