Clean up NPC table, add completion & more filters
Puts a limit on the size of the Name and Quest columns, which helps
prevent those columns from taking up a bunch of space and forcing the
rewards to wrap.

Adds NPC completion, allowing users to track NPCs that they have
defeated. This is different from the old completion logic, which was
based on claiming all of an NPCs rewards. The progress bar now works
off the new logic, and the old logic is available as a filter.

Adds an "Update Defeated" button which magically determines a subset of
NPCs you have defeated by checking your collection for their exclusive
cards.