Localize card names and descriptions
Add  the traco gem which allows for easy localization with the addition
of columns suffixed with the locale name. Adds the new localization
columns for card name & description, and updates the card creation task.
Also adds the option to pass a language to the API for retrieving card
data for a specific language.