Localize NPC locations
Creates a new model for NPC locations which include the location name
and region. Updates the NPC task to create these regions as necessary.
Updates the NPC pages to use this new model. Adds region to NPC API
data.

Fixes the NPC task to skip creating rules when attempting to update
an existing NPC. Rule changes will need to be made by hand.