Add remaining Deck pages
Adds the ability to create, edit, update, and delete decks, AKA the rest
of the freaking owl. Adds a new gem for drag & drop to allow for easy
card reordering. Adds seed data for creating varied decks in bulk.
Updates the Deck index page with pagination and filters, as well as
buttons to take the user to their own set of decks or create a new deck.
Satisfies my CoffeeScript quota for the year.